+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - Issue \#1605: Fix ensures that STYLES parameter in GFI URL are correctly reset. 
 - WMS-Time: The `dimensionName` attribute is now being processed correctly again, so queries for individual dimensions are working again.
 - Fix package-lock.json for node version 24.15.0.
+- Print: Fixed print mask disappearing after toggling a layer via LayerPills or switching the base layer via BaseLayerSwitcher.
 
 ---